Force re-read of newly added pages to build
This generates a couple of build warnings, because some references are
missing on the first pass. This is expected, but we should either
silence these errors or defer resolution of these references until later.

This also doesn't include badges for similar reasons: because the references
in the badges can't be resolved in the first pass, the badges are not
linked to any pages.